+++ R/cran_comments.R
@@ -1,7 +1,7 @@
 info <- function(session = NULL) {
     if (is.null(session)) 
         session <- utils::sessionInfo()
-    if (identical(class(session), "sessionInfo"))
+    if (!identical(class(session), "sessionInfo"))
         warning("Argument session is not of class `sessionInfo`!")
     info <- c(session[["R.version"]][["version.string"]],
               paste0("Platform: ", session[["platform"]]),
@@ -33,6 +33,32 @@ get_travis_info <- function(travis_session_info = NULL, path = ".") {
     }
     return(info)
 }
+
+# get_gitlab_info()
+# gitlab_token <- readLines(file.path("~", ".gitlab_private_token.txt"))
+# get_gitlab_info(path = ".", private_token = gitlab_token, httr::use_proxy("10.127.255.17", 8080))
+get_gitlab_info <- function(path = ".", private_token, ...) {
+    if (missing(private_token) || is.null(private_token)) {
+        warning("You need a private token to access gitlab.")
+        info <- NULL
+    } else {
+        url <- get_git_url(get_remote_url(path))
+        project <- basename(url)
+        user <- tolower(basename(dirname(url)))
+        log <- get_gitlab_log(user = user, project = project, private_token, ...)
+        info <- eval_from_log(file = log, pattern = "=== packager info:")
+        info <- info(info)
+        rcmdcheck <- eval_from_log(log, pattern = "=== packager rcmdcheck:")
+        status <- grep("^Status", 
+                       strsplit(rcmdcheck$output$stdout, split = "\n")[[1]], 
+                       value = TRUE)
+        info <- c(info, status)
+    }
+    
+
+    return(info)
+}
+
 #' Provide a Template for Your Comments To CRAN
 #'
 #' Devtools' \code{\link{release}} reads a file \emph{cran-comments.md}. This
@@ -56,26 +82,44 @@ get_travis_info <- function(travis_session_info = NULL, path = ".") {
 #' set up to use \url{https://github.com/travis-ci/travis.rb}.
 #' @param name The name to sign with, if NA, the given name of the package
 #' maintainer as stated in file DESCRIPTION is used.
-#' @write_to_file Do write the comment to \emph{cran-comment.md}?
+#' @param write_to_file Do write the comment to \emph{cran-comment.md}?
+#' @param private_token A private token to access gitlab.com.
+#' @param proxy A proxy to use.
 #' @note By default this function writes to disk as side effect.
 #' @return Character vector containing the cran comments, which are written to
 #' cran-comments.md (see Note).
 #' @export
 #' @examples
-#' cat(provide_cran_comments(check_log = system.file("files", "check.Rout", 
-#'                                                     package = "packager"),
-#'                             travis_session_info = system.file("files", 
-#'                                                               "travis.log", 
-#'                                                               package =
-#'                                                                   "packager"),
-#'                             write_to_file = FALSE),
-#'     sep = "")
+#' gitlab_token <- NULL
+#' \dontrun{
+#' gitlab_token <- readLines(file.path("~", ".gitlab_private_token.txt"))
+#' }
+#' 
+#' check_log <- system.file("files", "check.Rout", package = "packager")
+#' travis_session_info <- system.file("files", "travis.log", package = "packager")
+#' if (Sys.info()[["nodename"]] == "fvafrdebianCU") {
+#'     comments <- provide_cran_comments(path = ".",
+#'                                       check_log = check_log,
+#'                                       travis_session_info = travis_session_info,
+#'                                       write_to_file = TRUE, 
+#'                                       private_token = gitlab_token,
+#'                                       proxy = httr::use_proxy("10.127.255.17", 8080))
+#' 
+#' } else {
+#'     comments <- provide_cran_comments(path = ".",
+#'                                       check_log = check_log,
+#'                                       travis_session_info = travis_session_info,
+#'                                       write_to_file = FALSE, 
+#'                                       private_token = gitlab_token)
+#' }
+#' cat(comments, sep = "")
 provide_cran_comments <- function(check_log = NULL,
                                   path = ".",
                                   initial = FALSE,
                                   travis_session_info = NULL,
                                   write_to_file = TRUE,
-                                  name = NA) {
+                                  private_token = NULL,
+                                  name = NA, proxy = NULL) {
     if (is.na(name)) {
         name <- tryCatch({
             maintainer <- desc::desc_get_author(role = "cre", file = path)
@@ -122,7 +166,14 @@ provide_cran_comments <- function(check_log = NULL,
     comments <- c(comments, "- ", paste(here, collapse = "\n    "), "\n")
     travis_info <- get_travis_info(travis_session_info, path) 
     if (!is.null(travis_info)) 
-        comments <- c(comments, "- ", paste(travis_info, collapse = "\n    "), "\n")
+        comments <- c(comments, paste(c("- travis-ci.org", travis_info), collapse = "\n  "), "\n")
+    if (is.null(proxy)) {
+        gitlab_info <- get_gitlab_info(path = path, private_token = private_token, proxy) 
+    } else {
+        gitlab_info <- get_gitlab_info(path = path, private_token = private_token, proxy) 
+    }
+    if (!is.null(gitlab_info)) 
+        comments <- c(comments, paste(c("- gitlab.com", gitlab_info), collapse = "\n  "), "\n")
     comments <- c(comments, "- win-builder (devel)", "\n")
     if (! as.logical(file.access(".", mode = 2))) # see ?file.acces return/note
         if (isTRUE(write_to_file))