+++ NEWS.md
@@ -1,5 +1,6 @@
 # packager 0.15.3.9000
 
+* Added functions dput and evaluate objects via log file.
 * Added function get\_gitlab\_log() to read a check log from gitlab.
 
 # packager 0.15.3