+++ NEWS.md
@@ -1,3 +1,7 @@
+# packager 0.16.0.9000
+
+* FIXME
+
 # packager 0.16.0
 
 * provide\_cran\_comments() now reads info from logs on gitlab.com, given that